+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ master
 - Remove nan values from met data
 - Added backup sync command for syncronising the backups with the external drives
     which are removed from the vessel after the voyage
+- Added scripts for removing and reenabling the external harddrives on the aurora
 
 0.1.1
 -----